# shared/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/schedule/": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "Get current schedule",
        "description": "Retrieves the full schedule using the GetSchedule interactor.",
        "operationId": "get_schedule_schedule__get",
        "responses": {
          "200": {
            "description": "Schedule retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScheduleOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedule/changes/": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "List schedule audit log",
        "description": "Returns a history of all modifications made to the schedule, including skips, moves, and status changes.",
        "operationId": "list_schedule_changes_schedule_changes__get",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule changes retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListScheduleChangesResult"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/changes/{schedule_change_id}": {
      "delete": {
        "tags": [
          "Schedule"
        ],
        "summary": "Undo a specific schedule change",
        "description": "Reverts a previously made change to the schedule using its unique change ID.",
        "operationId": "undo_schedule_change_schedule_changes__schedule_change_id__delete",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Schedule change ID to undo.",
              "title": "Schedule Change Id"
            },
            "description": "Schedule change ID to undo."
          }
        ],
        "responses": {
          "204": {
            "description": "Change successfully undone."
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Schedule change ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorMessage": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorMessage"
      },
      "GetScheduleOutput": {
        "properties": {
          "schedule": {
            "items": {
              "$ref": "#/components/schemas/ScheduleEventFullDTO"
            },
            "type": "array",
            "title": "Schedule"
          }
        },
        "type": "object",
        "required": [
          "schedule"
        ],
        "title": "GetScheduleOutput"
      },
...
      "ListScheduleChangesResult": {
        "properties": {
          "schedule_changes": {
            "items": {
              "$ref": "#/components/schemas/ScheduleChangeFullDTO"
            },
            "type": "array",
            "title": "Schedule Changes"
          }
        },
        "type": "object",
        "required": [
          "schedule_changes"
        ],
        "title": "ListScheduleChangesResult"
      },
...
      "ScheduleChangeEventDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "order": {
            "type": "number",
            "title": "Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "title",
          "order"
        ],
        "title": "ScheduleChangeEventDTO"
      },
      "ScheduleChangeFullDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/ScheduleChangeType"
          },
          "mailing_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "next_event_changed": {
            "type": "boolean",
            "title": "Next Event Changed"
          },
          "changed_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleChangeEventDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "argument_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleChangeEventDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleChangeUserDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "mailing_id",
          "user_id",
          "next_event_changed",
          "changed_event",
          "argument_event",
          "user"
        ],
        "title": "ScheduleChangeFullDTO"
      },
      "ScheduleChangeType": {
        "type": "string",
        "enum": [
          "set_as_current",
          "moved",
          "skipped",
          "unskipped"
        ],
        "title": "ScheduleChangeType"
      },
      "ScheduleChangeUserDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "ScheduleChangeUserDTO"
      },
      "ScheduleEventFullDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "order": {
            "type": "number",
            "title": "Order"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current"
          },
          "is_skipped": {
            "type": "boolean",
            "title": "Is Skipped"
          },
          "nomination_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nomination Title"
          },
          "block_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Title"
          },
          "queue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "time_until": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Until"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "title",
          "duration",
          "order",
          "is_current",
          "is_skipped",
          "nomination_title",
          "block_title",
          "queue",
          "time_until"
        ],
        "title": "ScheduleEventFullDTO"
      },
...
      "ValidationErrorDetail": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "type"
        ],
        "title": "ValidationErrorDetail"
      },
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "VALIDATION_ERROR"
          },
          "details": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ValidationErrorDetail"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "title": "ValidationErrorResponse"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Session cookie set after login.",
        "in": "cookie",
        "name": "session_id"
      }
    }
  }
}